chore: Update Angular and Bootstrap dependencies
This commit updates the Angular and Bootstrap dependencies in the project. The Angular dependencies in the `angular.json` and `package.json` files have been updated to version 17.3.0, and the Bootstrap dependency in the `package.json` file has been updated to version 5.3.3. Additionally, the Bootstrap CSS and JS files have been added to the `styles` and `scripts` arrays in the `angular.json` file, respectively.

The purpose of this update is to ensure that the project is using the latest stable versions of Angular and Bootstrap, which may include bug fixes, performance improvements, and new features.

Please note that this commit does not include any functional changes to the codebase, but rather focuses on updating the project's dependencies.

```